CommonSettingsView is now CommonConfigView. Contains only commonly used controls useful to have in the main window.
Added ConfigView class which holds all config controls. It contains a CommonConfigView
plus extra options less commonly needed.

Added a ConfigDialog class which shows a ConfigView in a modal dialog.

Added a configure action which opens a ConfigDialog when triggered. Can be triggered
from the new Settings menu in the menubar, or from the system tray context menu.

Various other minor modifications to support the above.